+++ extensions/validator/src/commands/mod.rs
@@ -1,7 +1,7 @@
 use strum::{EnumDiscriminants, EnumIter, EnumMessage};
 
-pub mod proposals;
-pub mod validators;
+mod proposals;
+mod validators;
 
 #[derive(Debug, Clone, EnumDiscriminants, interactive_clap_derive::InteractiveClap)]
 #[strum_discriminants(derive(EnumMessage, EnumIter))]
@@ -10,12 +10,12 @@ pub mod validators;
 pub enum TopLevelCommand {
     #[strum_discriminants(strum(message = "Lookup validators for given epoch"))]
     /// Lookup validators for given epoch
-    Validators(self::validators::operation_mode::OperationMode),
+    Validators(self::validators::OperationMode),
     #[strum_discriminants(strum(
         message = "Show both new proposals in the current epoch as well as current validators who are implicitly proposing"
     ))]
     /// Show both new proposals in the current epoch as well as current validators who are implicitly proposing
-    Proposals(self::proposals::operation_mode::OperationMode),
+    Proposals(self::proposals::OperationMode),
 }
 
 impl TopLevelCommand {
