+++ extensions/validator/src/commands/proposals.rs
@@ -1,6 +1,22 @@
+use dialoguer::Input;
 use strum::{EnumDiscriminants, EnumIter, EnumMessage};
 
-pub mod server;
+use crate::common::display_proposals_info;
+
+#[derive(Debug, Clone, interactive_clap_derive::InteractiveClap)]
+#[interactive_clap(context = ())]
+pub struct OperationMode {
+    #[interactive_clap(named_arg)]
+    /// Prepare and, optionally, submit a new transaction with online mode
+    pub network: NetworkArgs,
+}
+
+#[derive(Debug, Clone, interactive_clap_derive::InteractiveClap)]
+#[interactive_clap(context = ())]
+pub struct NetworkArgs {
+    #[interactive_clap(subcommand)]
+    selected_server: SelectServer,
+}
 
 #[derive(Debug, Clone, EnumDiscriminants, interactive_clap_derive::InteractiveClap)]
 #[strum_discriminants(derive(EnumMessage, EnumIter))]
@@ -10,16 +26,16 @@ pub mod server;
 pub enum SelectServer {
     /// Provide data for the server https://rpc.testnet.near.org
     #[strum_discriminants(strum(message = "Testnet"))]
-    Testnet(self::server::Server),
+    Testnet(Server),
     /// Provide data for the server https://rpc.mainnet.near.org
     #[strum_discriminants(strum(message = "Mainnet"))]
-    Mainnet(self::server::Server),
+    Mainnet(Server),
     /// Provide data for the server https://rpc.betanet.near.org
     #[strum_discriminants(strum(message = "Betanet"))]
-    Betanet(self::server::Server),
+    Betanet(Server),
     /// Provide data for a manually specified server
     #[strum_discriminants(strum(message = "Custom"))]
-    Custom(self::server::CustomServer),
+    Custom(CustomServer),
 }
 
 #[derive(Clone)]
@@ -57,6 +73,63 @@ impl From<SelectServerContext> for NetworkContext {
     }
 }
 
+#[derive(Debug, Clone, interactive_clap_derive::InteractiveClap)]
+#[interactive_clap(context = NetworkContext)]
+pub struct Server {}
+
+#[derive(Debug, Clone, interactive_clap_derive::InteractiveClap)]
+#[interactive_clap(input_context = SelectServerContext)]
+#[interactive_clap(output_context = NetworkContext)]
+pub struct CustomServer {
+    #[interactive_clap(long)]
+    pub url: crate::common::AvailableRpcServerUrl,
+}
+
+struct CustomServerContext {
+    pub url: crate::common::AvailableRpcServerUrl,
+}
+
+impl CustomServerContext {
+    fn from_previous_context(
+        _previous_context: SelectServerContext,
+        scope: &<CustomServer as interactive_clap::ToInteractiveClapContextScope>::InteractiveClapContextScope,
+    ) -> Self {
+        Self {
+            url: scope.url.clone(),
+        }
+    }
+}
+
+impl From<CustomServerContext> for NetworkContext {
+    fn from(item: CustomServerContext) -> Self {
+        Self {
+            connection_config: crate::common::ConnectionConfig::from_custom_url(&item.url),
+        }
+    }
+}
+
+impl CustomServer {
+    pub fn input_url(
+        _context: &SelectServerContext,
+    ) -> color_eyre::eyre::Result<crate::common::AvailableRpcServerUrl> {
+        Ok(Input::new()
+            .with_prompt("What is the RPC endpoint?")
+            .interact_text()?)
+    }
+}
+
+impl OperationMode {
+    pub async fn process(self) -> crate::CliResult {
+        self.network.process().await
+    }
+}
+
+impl NetworkArgs {
+    pub async fn process(self) -> crate::CliResult {
+        self.selected_server.process().await
+    }
+}
+
 impl SelectServer {
     pub async fn process(self) -> crate::CliResult {
         Ok(match self {
@@ -76,5 +149,23 @@ impl SelectServer {
                 custom_server.process().await?;
             }
         })
+    }
+}
+
+impl Server {
+    pub async fn process(
+        self,
+        connection_config: crate::common::ConnectionConfig,
+    ) -> crate::CliResult {
+        display_proposals_info(&connection_config).await?;
+        Ok(())
+    }
+}
+
+impl CustomServer {
+    pub async fn process(self) -> crate::CliResult {
+        let connection_config = crate::common::ConnectionConfig::from_custom_url(&self.url);
+        display_proposals_info(&connection_config).await?;
+        Ok(())
     }
 }
