--- extensions/validator/src/commands/proposals/mod.rs
@@ -1,1 +1,0 @@
-pub mod operation_mode;
