--- extensions/validator/src/commands/validators/mod.rs
@@ -1,3 +1,0 @@
-pub mod block_id;
-pub mod epoch;
-pub mod operation_mode;
